cifar10-cnn-classifier/requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Machine Learning
scikit-learn>=1.0.0

# Progress Bars
tqdm>=4.64.0

# Configuration
PyYAML>=6.0

# Optional: For advanced features
# tensorboard>=2.8.0
# wandb>=0.12.0
# pillow>=8.3.0